requirements.txt
streamlit>=1.37.0
torch>=2.2.0
diffusers>=0.31.0
transformers>=4.43.0
accelerate>=0.30.0
safetensors>=0.4.3
imageio>=2.34.0
Pillow>=10.0.0
